// src/app/modules/staff/assignments/model/class-home-work-model.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { FileBase64Model } from './file-base64-model';


/**
 * 
 */
export interface ClassHomeWorkModel { 
     /**
     * 
     */
    id?: string;
    /**
     * 
     */
    boardId?: number;
    /**
     * 
     */
    schoolAcademicSessionClassSectionIds?: Array<number>;
    /**
     * 
     */
    typeId?: number;
    /**
     * 
     */
    courseId?: number;
    /**
     * 
     */
    homeworkName?: string;
    /**
     * 
     */
    homeworkDescription?: string;
    /**
     * Gets or Sets
     */
    file?: FileBase64Model;
    /**
     * 
     */
    isPostToCalendar?: boolean;
    /**
     * 
     */
    issueDate?: Date;
    /**
     * 
     */
    dueDate?: Date;
    /**
     * 
     */
    isGraded?: boolean;
    /**
     * 
     */
    testTypeId?: number;
    /**
     * 
     */
    gradeTermId?: number;
    /**
     * 
     */
    maxMarks?: number;
    /**
     * 
     */
    testId?: number;
}